+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -7,7 +7,6 @@ import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.dto.ResponseBookingDto;
 import ru.practicum.shareit.booking.enums.BookingState;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -18,31 +17,31 @@ public class BookingController {
 
     @PostMapping
     public ResponseBookingDto addBookingRequest(@Valid @RequestBody BookingDto bookingDto,
-                                                @RequestHeader("X-Sharer-User-Id") BigInteger userId) {
+                                                @RequestHeader("X-Sharer-User-Id") Long userId) {
         return bookingService.addBookingRequest(bookingDto, userId);
     }
 
     @PatchMapping("/{bookingId}")
-    public ResponseBookingDto approveBooking(@PathVariable BigInteger bookingId,
-                                             @RequestHeader("X-Sharer-User-Id") BigInteger userId,
+    public ResponseBookingDto approveBooking(@PathVariable Long bookingId,
+                                             @RequestHeader("X-Sharer-User-Id") Long userId,
                                              @RequestParam Boolean approved) {
         return bookingService.approveBooking(bookingId, userId, approved);
     }
 
     @GetMapping("/{bookingId}")
-    public ResponseBookingDto getBooking(@PathVariable BigInteger bookingId,
-                                         @RequestHeader("X-Sharer-User-Id") BigInteger userId) {
+    public ResponseBookingDto getBooking(@PathVariable Long bookingId,
+                                         @RequestHeader("X-Sharer-User-Id") Long userId) {
         return bookingService.getBooking(bookingId, userId);
     }
 
     @GetMapping
-    public List<ResponseBookingDto> getBookings(@RequestHeader("X-Sharer-User-Id") BigInteger userId,
+    public List<ResponseBookingDto> getBookings(@RequestHeader("X-Sharer-User-Id") Long userId,
                                                 @RequestParam(defaultValue = "ALL") BookingState state) {
         return bookingService.getBookings(userId, state);
     }
 
     @GetMapping("/owner")
-    public List<ResponseBookingDto> getBookingsOfItemsOfOwner(@RequestHeader("X-Sharer-User-Id") BigInteger userId,
+    public List<ResponseBookingDto> getBookingsOfItemsOfOwner(@RequestHeader("X-Sharer-User-Id") Long userId,
                                                               @RequestParam(defaultValue = "ALL") BookingState state) {
         return bookingService.getBookingsOfAllItemsOfOwner(userId, state);
     }
