+++ src/main/java/ru/practicum/shareit/booking/BookingService.java
@@ -4,17 +4,16 @@ import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.dto.ResponseBookingDto;
 import ru.practicum.shareit.booking.enums.BookingState;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface BookingService {
-    ResponseBookingDto addBookingRequest(BookingDto bookingDto, BigInteger userId);
+    ResponseBookingDto addBookingRequest(BookingDto bookingDto, Long userId);
 
-    ResponseBookingDto approveBooking(BigInteger bookingId, BigInteger userId, Boolean approved);
+    ResponseBookingDto approveBooking(Long bookingId, Long userId, Boolean approved);
 
-    ResponseBookingDto getBooking(BigInteger bookingId, BigInteger userId);
+    ResponseBookingDto getBooking(Long bookingId, Long userId);
 
-    List<ResponseBookingDto> getBookings(BigInteger userId, BookingState state);
+    List<ResponseBookingDto> getBookings(Long userId, BookingState state);
 
-    List<ResponseBookingDto> getBookingsOfAllItemsOfOwner(BigInteger userId, BookingState state);
+    List<ResponseBookingDto> getBookingsOfAllItemsOfOwner(Long userId, BookingState state);
 }
