+++ src/main/java/ru/practicum/shareit/item/CommentStorage.java
@@ -4,10 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ru.practicum.shareit.item.model.Comment;
 
-import java.math.BigInteger;
 import java.util.List;
 
-public interface CommentStorage extends JpaRepository<Comment, BigInteger> {
+public interface CommentStorage extends JpaRepository<Comment, Long> {
     @Query(value = "select com from Comment as com where com.item.id = ?1")
-    List<Comment> findAllByItemId(BigInteger itemId);
+    List<Comment> findAllByItemId(Long itemId);
 }
