+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -9,7 +9,6 @@ import ru.practicum.shareit.item.dto.CommentDto;
 import ru.practicum.shareit.item.dto.CreateCommentDto;
 import ru.practicum.shareit.item.dto.ItemDto;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -20,23 +19,23 @@ public class ItemController {
     private final ItemService itemService;
 
     @PostMapping
-    public ItemDto addItem(@Valid @RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") BigInteger userId) {
+    public ItemDto addItem(@Valid @RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") Long userId) {
         return itemService.addItem(itemDto, userId);
     }
 
     @PatchMapping("/{itemId}")
-    public ItemDto updateItem(@RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") BigInteger userId,
-                              @PathVariable BigInteger itemId) {
+    public ItemDto updateItem(@RequestBody ItemDto itemDto, @RequestHeader("X-Sharer-User-Id") Long userId,
+                              @PathVariable Long itemId) {
         return itemService.updateItem(itemDto, userId, itemId);
     }
 
     @GetMapping("/{itemId}")
-    public ItemDto getItemById(@PathVariable BigInteger itemId) {
+    public ItemDto getItemById(@PathVariable Long itemId) {
         return itemService.getItemById(itemId);
     }
 
     @GetMapping
-    public List<ItemDto> getItemsOfUser(@RequestHeader("X-Sharer-User-Id") BigInteger userId) {
+    public List<ItemDto> getItemsOfUser(@RequestHeader("X-Sharer-User-Id") Long userId) {
         return itemService.getItemsOfUser(userId);
     }
 
@@ -46,8 +45,8 @@ public class ItemController {
     }
 
     @PostMapping("/{itemId}/comment")
-    public CommentDto addComment(@RequestBody CreateCommentDto createCommentDto, @PathVariable BigInteger itemId,
-                                 @RequestHeader("X-Sharer-User-Id") BigInteger userId) {
+    public CommentDto addComment(@RequestBody CreateCommentDto createCommentDto, @PathVariable Long itemId,
+                                 @RequestHeader("X-Sharer-User-Id") Long userId) {
         log.info("Получили text = {}, itemId = {}, userId = {}", createCommentDto.getText(), itemId, userId);
         return itemService.addComment(createCommentDto, itemId, userId);
     }
