+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -6,8 +6,6 @@ import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.user.dto.UpdateUserDto;
 import ru.practicum.shareit.user.dto.UserDto;
 
-import java.math.BigInteger;
-
 @RestController
 @RequestMapping(path = "/users")
 @RequiredArgsConstructor
@@ -15,7 +13,7 @@ public class UserController {
     private final UserService userService;
 
     @GetMapping("/{userId}")
-    public UserDto getUserById(@PathVariable BigInteger userId) {
+    public UserDto getUserById(@PathVariable Long userId) {
         return userService.getUserById(userId);
     }
 
@@ -25,12 +23,12 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public UserDto updateUser(@Valid @RequestBody UpdateUserDto userDto, @PathVariable BigInteger userId) {
+    public UserDto updateUser(@Valid @RequestBody UpdateUserDto userDto, @PathVariable Long userId) {
         return userService.updateUser(userDto, userId);
     }
 
     @DeleteMapping("/{userId}")
-    public void deleteUserById(@PathVariable BigInteger userId) {
+    public void deleteUserById(@PathVariable Long userId) {
         userService.deleteUserById(userId);
     }
 }
