+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -3,14 +3,12 @@ package ru.practicum.shareit.user;
 import ru.practicum.shareit.user.dto.UpdateUserDto;
 import ru.practicum.shareit.user.dto.UserDto;
 
-import java.math.BigInteger;
-
 public interface UserService {
-    UserDto getUserById(BigInteger userId);
+    UserDto getUserById(Long userId);
 
     UserDto createUser(UserDto userDto);
 
-    UserDto updateUser(UpdateUserDto userDto, BigInteger userId);
+    UserDto updateUser(UpdateUserDto userDto, Long userId);
 
-    void deleteUserById(BigInteger userId);
+    void deleteUserById(Long userId);
 }
