+++ src/main/java/ru/practicum/shareit/user/UserStorage.java
@@ -2,8 +2,6 @@ package ru.practicum.shareit.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface UserStorage extends JpaRepository<User, BigInteger> {
+public interface UserStorage extends JpaRepository<User, Long> {
     boolean existsByEmail(String email);
 }
